requirements.txt
python-decouple
influxdb-client
pytz
board
opencv-python
spidev
adafruit-circuitpython-dht
numpy